requirements.txt
Pandas
numpy
Seaborn
matplotlib
scipy
scikit-learn
PyTorch
